pyproject.toml
[project]
name = "vibehack"
version = "1.7.0"
description = "Raw Shell / HitL Agentic Security Co-Pilot"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Proprietary"}
dependencies = [
    "typer>=0.12.0",
    "rich>=13.7.0",
    "pydantic>=2.7.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1",
    "litellm>=1.0.0",
    "fpdf2>=2.7.9",
    "playwright>=1.40.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
vibehack = "vibehack.cli:app"
vibehack-browser = "vibehack.toolkit.browser_agent:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

